feat: display unknown contract changes
